Extract copy functionality to utils
