+++ superset-frontend/src/views/CRUD/data/savedquery/SavedQueryList.tsx
@@ -41,6 +41,7 @@ import { IconName } from 'src/components/Icon';
 import { commonMenuData } from 'src/views/CRUD/data/common';
 import { SavedQueryObject } from 'src/views/CRUD/types';
 import SavedQueryPreviewModal from './SavedQueryPreviewModal';
+import copyTextToClipboard from '../../../../utils/copy';
 
 const PAGE_SIZE = 25;
 
@@ -154,39 +155,13 @@ function SavedQueryList({
 
   const copyQueryLink = useCallback(
     (id: number) => {
-      const selection: Selection | null = document.getSelection();
 
-      if (selection) {
-        selection.removeAllRanges();
-        const range = document.createRange();
-        const span = document.createElement('span');
-        span.textContent = `${window.location.origin}/superset/sqllab?savedQueryId=${id}`;
-        span.style.position = 'fixed';
-        span.style.top = '0';
-        span.style.clip = 'rect(0, 0, 0, 0)';
-        span.style.whiteSpace = 'pre';
-
-        document.body.appendChild(span);
-        range.selectNode(span);
-        selection.addRange(range);
-
-        try {
-          if (!document.execCommand('copy')) {
-            throw new Error(t('Not successful'));
-          }
-        } catch (err) {
-          addDangerToast(t('Sorry, your browser does not support copying.'));
-        }
-
-        document.body.removeChild(span);
-        if (selection.removeRange) {
-          selection.removeRange(range);
-        } else {
-          selection.removeAllRanges();
-        }
-
+      copyTextToClipboard(`${window.location.origin}/superset/sqllab?savedQueryId=${id}`)
+      .then(() => {
         addSuccessToast(t('Link Copied!'));
-      }
+      }).catch(() => {
+        addDangerToast(t('Sorry, your browser does not support copying.'));
+      })
     },
     [addDangerToast, addSuccessToast],
   );
